Disable build of GObject Introspection, fix cross compile
Not trivial in a cross setting as BUILD and HOST dev packages
must be installed side by side and the upstream debian ones
doesn't support it.

Confirmed that upstream python-gst-1.0, that depends on
gir1.2-gstreamer-1.0, can be installed and used alongside
our binary debs so we're not losing language bindings.

This disables plugins whose deps can't be installed in a cross
environment due to how they're packaged. They're not of great
importance to us though.

Change-Id: Id4169e96453193cc0cbbe50e262d5ab6d7f6180c